fix(compress): dequantize BITPIX=-64 tiles in full f64 precision (bit-for-bit CFITSIO/astropy parity) (#36)
* fix(compress): dequantize BITPIX=-64 tiles in full f64 precision, matching CFITSIO/astropy bit-for-bit

dither.unquantize/unquantizeNext computed the (stored - r + 0.5) * ZSCALE
+ ZZERO reconstruction in f64 but returned f32, and the tiled read loop
stored through f32 before the output cast - so every pixel of every
quantized (lossy) double image was silently truncated to a 24-bit
mantissa (~6e-5 absolute at values ~1000; 0/4096 pixels matched
funpack/astropy). Widen the return path to f64; ZBITPIX=-32 decodes are
byte-identical (same single f64->f32 rounding, moved to the end), which
the existing funpack bit-pattern goldens confirm.

The final "* ZSCALE + ZZERO" now uses @mulAdd: CFITSIO's and astropy's
C builds on FMA hardware contract that expression (clang/gcc default
-ffp-contract=on), so a plain mul-then-add still differed by 1 ULP on
~0.6% of double pixels. @mulAdd is a true fma on every Zig target
(software fallback included), giving byte-deterministic output that
matches the contracted reference builds - verified 0/4096 bitwise
diffs vs funpack AND astropy 8.0.1, in both directions (fpack-authored
file read by zigfitsio; zigfitsio-authored file decoded by funpack).

Coverage that would have caught this: a funpack-authored quantized
BITPIX=-64 golden (compress/tile_rice_ddith*, values ~1000 where the
f32 grid towers over the dither term) pinned via TiledImage.readAll(f64)
u64-bit equality, plus a write/read canary asserting the dequantized
pixels carry beyond-f32 precision.

Bug-hunt 2026-07-06 item 41.


* test(interop): cross-validate the quantized-double golden with astropy at <= 1 ULP

Astropy joins funpack and zigfitsio on the tile_rice_ddith pair. The
comparison is 1-ULP-tolerant, not bit-exact: the final "* ZSCALE +
ZZERO" is an FP-contraction point on which CFITSIO's own builds
disagree at f64 width (FMA-contracted arm64 vs non-contracted baseline
x86-64), so an astropy wheel built without FMA legitimately lands 1 ULP
away on a small fraction of pixels. zigfitsio's own decode stays pinned
bit-exactly in test/golden.zig via the byte-deterministic @mulAdd.


* test(compress): guard the f64 canary against a silent lossless fallback; doc accuracy

The canary's source field is itself not f32-representable, so a tile
that silently fell back to lossless storage (out == src verbatim) would
pass the beyond-f32 check without the dequantization path ever running.
Assert quantization was actually lossy (some pixel differs) so the
canary cannot go vacuous.

Also correct the measured FMA-divergence rate in the unquantize doc
(26/4096 = ~0.6%, not ~0.5%) and note the wasm32 software-fma decode
cost that byte-determinism buys.


---------

+++ src/compress/quantize.zig
@@ -547,7 +547,7 @@ test "quantizeTile: SUBTRACTIVE_DITHER_1 round-trips within delta via the matchi
     // Decode with the file-format reconstruction (same draws): |err| <= delta/2.
     var cur = dither.Dither.init(table, .subtractive_1, 42, 3);
     for (data, idata) |v, s| {
-        const rec = cur.unquantizeNext(s, p.bscale, p.bzero, std.math.nan(f32));
+        const rec = cur.unquantizeNext(s, p.bscale, p.bzero, std.math.nan(f64));
         try testing.expect(@abs(rec - v) <= 0.25 + 1e-6);
     }
 }
@@ -581,9 +581,9 @@ test "quantizeTile: SUBTRACTIVE_DITHER_2 preserves exact zeros via ZERO_VALUE" {
     try testing.expectEqual(dither.zero_value, idata[17]);
     var cur = dither.Dither.init(table, .subtractive_2, 1, 0);
     for (data, idata) |v, s| {
-        const rec = cur.unquantizeNext(s, r.quantized.bscale, r.quantized.bzero, std.math.nan(f32));
+        const rec = cur.unquantizeNext(s, r.quantized.bscale, r.quantized.bzero, std.math.nan(f64));
         if (v == 0.0) {
-            try testing.expectEqual(@as(f32, 0.0), rec);
+            try testing.expectEqual(@as(f64, 0.0), rec);
         } else {
             try testing.expect(@abs(rec - v) <= 0.0625 + 1e-6);
         }
